requirements.txt
altair==4.2.0
matplotlib==3.5.2
pandas==1.4.4
Pillow==9.3.0
streamlit==1.16.0
streamlit_option_menu==0.3.2
xgboost==1.6.2
